Corrected all warnings except for lodepng.  Its a complex beast.
Former-commit-id: dc5e3d9f63ebf5a52a26be7836d34f92349b2c9e

+++ Sample/source/TerrainChunkRenderer.h
@@ -14,7 +14,7 @@ class TerrainChunkRenderer {
     {
 		GLuint vertexArrayBuffer;
 		GLuint vertexData;
-		int vertexCount;
+		size_t vertexCount;
 	public:
 		HotChunk();
 		~HotChunk();
